+++ Runtime/AssetFolderAttribute.cs
@@ -1,13 +1,13 @@
 using System.Diagnostics;
-using UnityEngine;
 
 namespace SaintsField
 {
     [Conditional("UNITY_EDITOR")]
     [System.AttributeUsage(System.AttributeTargets.Field | System.AttributeTargets.Property, Inherited = true, AllowMultiple = false)]
-    public class FlagsDropdownAttribute: PropertyAttribute, ISaintsAttribute
+    public class AssetFolderAttribute: FolderAttribute
     {
-        public SaintsAttributeType AttributeType => SaintsAttributeType.Field;
-        public string GroupBy => "";
+        public AssetFolderAttribute(string folder="Assets", string title="Choose a folder inside assets", string groupBy = "") : base(folder, title, groupBy)
+        {
+        }
     }
 }