+++ Runtime/FolderAttribute.cs
@@ -1,0 +1,27 @@
+using System.Diagnostics;
+using UnityEngine;
+
+namespace SaintsField
+{
+    [Conditional("UNITY_EDITOR")]
+    [System.AttributeUsage(System.AttributeTargets.Field | System.AttributeTargets.Property, Inherited = true, AllowMultiple = false)]
+    public class FolderAttribute: PropertyAttribute, ISaintsAttribute
+    {
+        public SaintsAttributeType AttributeType => SaintsAttributeType.Other;
+        public string GroupBy { get; }
+
+        public readonly string Folder;
+        public readonly string Title;
+
+
+        public FolderAttribute(string folder="", string title="", string groupBy = "")
+        {
+            GroupBy = groupBy;
+
+            Folder = folder;
+            Title = title;
+        }
+
+
+    }
+}